e2e: expose test por to allow test customizations
This change expose the service and pod port of JIG interface
used when developing new tests to test Load Balancers, such as CCM,
leaving the default values to 80.

That customization allow to reuse existing test lib while developing
load balancer tests to environments which does not support to bind pod
in privileged ports (80).